+++ remoc_macro/src/method.rs
@@ -768,7 +768,7 @@ impl TraitMethod {
                 let req = ::remoc::rtc::Req::#req_type(req_value);
 
                 let mut guard = match self.monitor.pre_call(&req).await {
-                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::DefaultGuard),
+                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::PassGuard),
                     ::remoc::rtc::monitor::CallDecision::Guard(guard) => guard,
                     ::remoc::rtc::monitor::CallDecision::Drop => return Err(::remoc::rtc::CallError::Dropped.into()),
                 };
@@ -838,7 +838,7 @@ impl TraitMethod {
                 let req = ::remoc::rtc::Req::#req_type(req_value);
 
                 let mut guard = match self.monitor.pre_call(&req).await {
-                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::DefaultGuard),
+                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::PassGuard),
                     ::remoc::rtc::monitor::CallDecision::Guard(guard) => guard,
                     ::remoc::rtc::monitor::CallDecision::Drop => {
                         return ::remoc::rtc::Call::ready(
@@ -923,7 +923,7 @@ impl TraitMethod {
                 let req = ::remoc::rtc::Req::#req_type(req_value);
 
                 let mut guard = match self.monitor.pre_call(&req).await {
-                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::DefaultGuard),
+                    ::remoc::rtc::monitor::CallDecision::Pass => ::std::boxed::Box::new(::remoc::rtc::monitor::PassGuard),
                     ::remoc::rtc::monitor::CallDecision::Guard(guard) => guard,
                     ::remoc::rtc::monitor::CallDecision::Drop => {
                         return ::remoc::rtc::Call::ready(#full_name, ::remoc::rtc::Response::from_call_error(